GEM-ML/WaterSegmentation/requirements.txt
absl-py==1.3.0
accumulation-tree==0.6.2
aenum==3.1.11
affine==2.3.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.94
botocore==1.27.94
branca==0.5.0
cachetools==5.2.0
captum==0.5.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.5
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
eo-learn-core==1.3.0
eo-learn-coregistration==1.3.0
eo-learn-features==1.3.0
eo-learn-geometry==1.3.0
eo-learn-io==1.3.0
eo-learn-mask==1.3.0
eo-learn-ml-tools==1.3.0
eo-learn-visualization==1.3.0
executing==1.1.1
fastjsonschema==2.16.2
Fiona==1.8.22
folium==0.13.0
fonttools==4.37.4
fs==2.4.16
fs-s3fs==1.1.1
geopandas==0.11.1
google-auth==2.13.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.50.0
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.1
ipython==8.5.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lightgbm==3.3.3
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
multiprocessing-on-dill==3.5.0a4
munch==2.5.0
mypy-extensions==0.4.3
natsort==8.2.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.5.1
notebook_shim==0.2.0
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
opencv-contrib-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
pretrainedmodels==0.7.4
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
pyproj==3.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.5
pyudorandom==1.0.0
PyWavelets==1.4.1
pyzmq==24.0.1
rasterio==1.3.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.0.1
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
segmentation-models-pytorch==0.3.0
Send2Trash==1.8.0
sentinelhub==3.8.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.3.2.post1
stack-data==0.5.1
tdigest==0.5.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.4.12
tinycss2==1.2.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
utm==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
zipp==3.9.0
